+++ src/sections/dataset/dataset-files/files-table/file-info/file-info-cell/file-info-data/FileTitle.tsx
@@ -1,6 +1,7 @@
 import { LinkToPage } from '../../../../../../shared/link-to-page/LinkToPage'
 import { Route } from '../../../../../../Route.enum'
 import { DatasetVersionNumber } from '../../../../../../../dataset/domain/models/Dataset'
+import { DvObjectType } from '../../../../../../../shared/hierarchy/domain/models/UpwardHierarchyNode'
 
 interface FileTitleProps {
   id: number
@@ -12,6 +13,7 @@ export function FileTitle({ id, name, datasetVersionNumber }: FileTitleProps) {
   return (
     <LinkToPage
       page={Route.FILES}
+      type={DvObjectType.FILE}
       searchParams={{ id: id.toString(), datasetVersion: datasetVersionNumber.toSearchParam() }}>
       {name}
     </LinkToPage>