+++ packages/design-system/CHANGELOG.md
@@ -46,6 +46,8 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - **FormTextArea:** extend Props Interface to accept `autoFocus` prop.
 - **FormSelect:** extend Props Interface to accept `autoFocus` prop.
 - **Stack:** NEW Stack element to manage layouts.
+- **TransferList:** NEW TransferList component to transfer items between two list, also sortable.
+- **Table:** extend Props Interface to accept `bordered` prop to add or remove borders on all sides of the table and cells. Defaults to true.
 
 # [1.1.0](https://github.com/IQSS/dataverse-frontend/compare/@iqss/dataverse-design-system@1.0.1...@iqss/dataverse-design-system@1.1.0) (2024-03-12)